Make enabling Auto Layout non-destructive, matching Figma
Turning Auto Layout on for a frame with existing children used to snap
everything to zero-gap/zero-padding immediately, since layoutMode changed
straight to Horizontal/Vertical with padding=0 and gap=0 already set, and
the frame was force-switched to Hug sizing. Figma instead infers padding
and gap from the current arrangement so the toggle is a no-op visually.

New AutoLayout.cpp helper InferAutoLayoutSpacing(Frame*, LayoutMode),
called once at the None -> enabled transition:
- Padding is inferred from the union bbox of the frame's current children
  vs. its own bounds (axis-independent, always computable).
- Gap is inferred from the spacing between the first two children sorted
  along the axis being enabled, but only when that spacing is actually
  non-overlapping; switching straight to an axis the children were never
  arranged along (e.g. a vertical stack toggled to Horizontal) falls back
  to Figma's own default of 10, since there's nothing sensible to infer.
- An empty frame gets 10px padding on all sides as a starting default,
  and its size is left completely alone: it's kept on Fixed sizing rather
  than switching to Hug, since an empty Hug frame collapses to just its
  padding the moment the next layout pass runs.

Also fixes a real correctness bug this surfaced: Frame's padding fields
were UInt8 (max 255), silently clamping any inferred (or manually typed)
padding above that -- a completely ordinary value on non-trivial designs.
Widened to UInt16 across the data model, serializer (bumped to v19, with
a backward-compatible 1-byte read path for older files), and the
Inspector's padding edit fields (clamp raised from 255 to 9999).

+++ src/core/Frame.h
@@ -41,10 +41,12 @@ public:
     UInt16       layoutGap            = 0;
     UInt16       layoutCounterGap     = 0;    // cross-axis gap between Wrap rows
     bool         layoutCounterGapAuto = false; // true = SpaceBetween distribution for rows
-    UInt8        paddingTop    = 0;
-    UInt8        paddingRight  = 0;
-    UInt8        paddingBottom = 0;
-    UInt8        paddingLeft   = 0;
+    // UInt16, not UInt8: inferred padding (see AutoLayout.cpp's InferAutoLayoutSpacing)
+    // routinely exceeds 255 on real designs.
+    UInt16       paddingTop    = 0;
+    UInt16       paddingRight  = 0;
+    UInt16       paddingBottom = 0;
+    UInt16       paddingLeft   = 0;
     PrimaryAlign primaryAlign  = PrimaryAlign::Start;
     CrossAlign   crossAlign    = CrossAlign::Start;
